requirements.txt
fpdf2==2.5.0
Pillow==9.0.1
